chore: Rename dotfiles to dotfile
